Move scenery group object
